Progress #15: inline EventConsumer into ZmqEventConsumer
